+++ run.sh
@@ -11,9 +11,10 @@ BRANCH=$(git rev-parse --abbrev-ref HEAD)
 GOLDFLAGS="-X main.Commit=${COMMIT} -X main.Branch=${BRANCH}"
 
 go run -ldflags "${GOLDFLAGS}" cmd/skycoin/skycoin.go \
-    --gui-dir="${DIR}/src/gui/static/" \
-    --launch-browser=true \
-    --enable-wallet-api=true \
+    -gui-dir="${DIR}/src/gui/static/" \
+    -launch-browser=true \
+    -enable-wallet-api=true \
+    -rpc-interface=false \
     $@
 
 popd >/dev/null