+++ src/gui/static/src/app/app.module.ts
@@ -32,12 +32,12 @@ import { HeaderComponent } from './components/layout/header/header.component';
 import { TopBarComponent } from './components/layout/header/top-bar/top-bar.component';
 import { FooterComponent } from './components/layout/footer/footer.component';
 import { PriceService } from './price.service';
-import { LoadWalletComponent } from './components/pages/wallets/load-wallet/load-wallet.component';
 import { TransactionListComponent } from './components/pages/transaction-list/transaction-list.component';
 import { TransactionDetailComponent } from './components/pages/transaction-list/transaction-detail/transaction-detail.component';
 import { NavBarComponent } from './components/layout/header/nav-bar/nav-bar.component';
 import { WalletDetailComponent } from './components/pages/wallets/wallet-detail/wallet-detail.component';
 import { ModalComponent } from './components/layout/modal/modal.component';
+import { PasswordDialogComponent } from './components/layout/password-dialog/password-dialog.component';
 import { MatButtonModule } from '@angular/material/button';
 import { MatCardModule } from '@angular/material/card';
 import { MatDialogModule } from '@angular/material/dialog';
@@ -54,9 +54,11 @@ import { MatSnackBarModule } from '@angular/material/snack-bar';
 import { MatTabsModule } from '@angular/material/tabs';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { MatTooltipModule } from '@angular/material/tooltip';
+import { MatCheckboxModule } from '@angular/material/checkbox';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { HttpClientModule } from '@angular/common/http';
 import { AppService } from './services/app.service';
+import { SeedModalComponent } from './components/pages/settings/backup/seed-modal/seed-modal.component';
 
 
 const ROUTES = [
@@ -123,7 +125,6 @@ const ROUTES = [
     ExplorerComponent,
     FooterComponent,
     HeaderComponent,
-    LoadWalletComponent,
     NetworkComponent,
     OutputsComponent,
     PendingTransactionsComponent,
@@ -138,15 +139,18 @@ const ROUTES = [
     NavBarComponent,
     WalletDetailComponent,
     ModalComponent,
+    PasswordDialogComponent,
+    SeedModalComponent,
   ],
   entryComponents: [
     AddDepositAddressComponent,
     CreateWalletComponent,
     ChangeNameComponent,
-    LoadWalletComponent,
     QrCodeComponent,
     SendSkycoinComponent,
     TransactionDetailComponent,
+    PasswordDialogComponent,
+    SeedModalComponent,
   ],
   imports: [
     BrowserModule,
@@ -168,6 +172,7 @@ const ROUTES = [
     MatTabsModule,
     MatToolbarModule,
     MatTooltipModule,
+    MatCheckboxModule,
     NgxDatatableModule,
     NoopAnimationsModule,
     ReactiveFormsModule,