+++ src/gui/wallet.go
@@ -4,6 +4,7 @@ package gui
 import (
 	"fmt"
 	"net/http"
+	"sort"
 	"strconv"
 
 	"github.com/skycoin/skycoin/src/cipher"
@@ -525,10 +526,15 @@ func walletTransactionsHandler(gateway Gatewayer) http.HandlerFunc {
 }
 
 // Returns all loaded wallets
-// RUI: /wallets
+// URI: /wallets
 // Method: GET
 func walletsHandler(gateway Gatewayer) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
+		if r.Method != http.MethodGet {
+			wh.Error405(w)
+			return
+		}
+
 		wlts, err := gateway.GetWallets()
 		if err != nil {
 			switch err {
@@ -540,7 +546,7 @@ func walletsHandler(gateway Gatewayer) http.HandlerFunc {
 			return
 		}
 
-		var wrs []*WalletResponse
+		wrs := make([]*WalletResponse, 0, len(wlts))
 		for _, wlt := range wlts {
 			wr, err := NewWalletResponse(wlt)
 			if err != nil {
@@ -550,6 +556,11 @@ func walletsHandler(gateway Gatewayer) http.HandlerFunc {
 
 			wrs = append(wrs, wr)
 		}
+
+		sort.Slice(wrs, func(i, j int) bool {
+			return wrs[i].Meta.Timestamp < wrs[j].Meta.Timestamp
+		})
+
 		wh.SendJSONOr500(logger, w, wrs)
 	}
 }